Add unit test for deployment
- `compilation_test.py` script only checks that every contract compiles correctly
- For each cairo contract you can create a python file following ERC20 example that compiles it, deploys it and checks the output of calling functions from it
- Modify makefile to add these last tests
