+++ Makefile
@@ -23,8 +23,9 @@ test_bats: warp $(BATS_FILES)
 .PHONY: test_bats
 
 test_yul: warp
-	python3.7 -m pytest scripts/yul/transpile_test.py -v --tb=short
-	python3.7 -m pytest scripts/yul/starknet_test.py -v --tb=short
+	python -m pytest scripts/yul/transpile_test.py -v --tb=short
+	python -m pytest scripts/yul/compilation_test.py -v --tb=short
+	python -m pytest tests/yul/ -v --tb=short
 .PHONY: test_yul
 
 $(BATS_DIR)/test-%.bats: \
