+++ tests/yul/ERC20_test.py
@@ -1,0 +1,33 @@
+import pytest
+import os
+
+from starknet.compile import compile_starknet_files
+from starkware.starknet.testing.starknet import Starknet
+from starkware.starknet.testing.contract import StarknetContract
+
+warp_root = os.path.abspath(os.path.join(__file__, "../../.."))
+test_dir = __file__
+
+@pytest.mark.asyncio
+async def test_starknet():
+    contract_file = test_dir[:-8] + '.cairo'
+    cairo_path = f"{warp_root}/warp/cairo-src"
+    contract_definition = compile_starknet_files([contract_file], debug_info=True, cairo_path=[cairo_path])
+
+    starknet = await Starknet.empty()
+    contract_address = await starknet.deploy(
+        contract_definition=contract_definition)
+    contract = StarknetContract(
+        starknet=starknet,
+        abi=contract_definition.abi,
+        contract_address=contract_address,
+    )
+
+    sender = 30
+    receiver = 20
+    assert await contract.fun_deposit_external(
+        var_sender_72_low=sender,
+        var_sender_72_high=0,
+        var_value_low=500,
+        var_value_high=0
+    ).call() == (21, 0, 12, 0)
