+++ tests/yul/if-flattening.sol
@@ -5,16 +5,14 @@ contract WARP {
 
     function transferFrom(uint src, uint dst, uint wad, uint sender)
         public payable
-        returns (uint)
-    {
-      uint res = 0;
-        if (src != sender) {
-            allowance[src][sender] -= wad;
-            res = 1;
-        } else {
-            res = 2;
-        }
-
-        return res;
+        returns (uint) {
+    uint res = 0;
+    if (src != sender) {
+      allowance[src][sender] -= wad;
+        res = 1;
+      } else {
+        res = 2;
+      }
+    return res;
     }
 }
